Fix Uint64 type errors in StoreChecks test data

+++ tests/consensus/devnet/fc/test_tick_system.py
@@ -14,6 +14,7 @@ from consensus_testing import (
 
 from lean_spec.subspecs.containers.slot import Slot
 from lean_spec.subspecs.containers.validator import ValidatorIndex
+from lean_spec.types.uint import Uint64
 
 pytestmark = pytest.mark.valid_until("Devnet")
 
@@ -52,7 +53,7 @@ def test_tick_interval_progression_through_full_slot(
             TickStep(
                 time=12,
                 checks=StoreChecks(
-                    time=15,
+                    time=Uint64(15),
                     head_slot=Slot(2),
                     head_root_label="block_2",
                 ),
@@ -62,7 +63,7 @@ def test_tick_interval_progression_through_full_slot(
             TickStep(
                 time=13,
                 checks=StoreChecks(
-                    time=16,
+                    time=Uint64(16),
                     head_slot=Slot(2),
                     head_root_label="block_2",
                 ),
@@ -73,7 +74,7 @@ def test_tick_interval_progression_through_full_slot(
             TickStep(
                 time=14,
                 checks=StoreChecks(
-                    time=17,
+                    time=Uint64(17),
                     head_slot=Slot(2),
                     head_root_label="block_2",
                 ),
@@ -118,7 +119,7 @@ def test_tick_interval_progression_through_full_slot(
             TickStep(
                 time=15,
                 checks=StoreChecks(
-                    time=18,
+                    time=Uint64(18),
                     head_slot=Slot(2),
                     head_root_label="block_2",
                     safe_target_slot=Slot(2),
@@ -152,7 +153,7 @@ def test_tick_interval_progression_through_full_slot(
             TickStep(
                 time=16,
                 checks=StoreChecks(
-                    time=20,
+                    time=Uint64(20),
                     head_slot=Slot(2),
                     head_root_label="block_2",
                     safe_target_slot=Slot(2),
@@ -198,7 +199,7 @@ def test_on_tick_advances_across_multiple_empty_slots(
             TickStep(
                 time=8,
                 checks=StoreChecks(
-                    time=10,
+                    time=Uint64(10),
                     head_slot=Slot(1),
                     head_root_label="block_1",
                 ),
@@ -207,7 +208,7 @@ def test_on_tick_advances_across_multiple_empty_slots(
             TickStep(
                 time=12,
                 checks=StoreChecks(
-                    time=15,
+                    time=Uint64(15),
                     head_slot=Slot(1),
                     head_root_label="block_1",
                 ),
@@ -216,7 +217,7 @@ def test_on_tick_advances_across_multiple_empty_slots(
             TickStep(
                 time=16,
                 checks=StoreChecks(
-                    time=20,
+                    time=Uint64(20),
                     head_slot=Slot(1),
                     head_root_label="block_1",
                 ),
@@ -252,7 +253,7 @@ def test_tick_interval_0_skips_acceptance_when_not_proposer(
             TickStep(
                 interval=14,
                 checks=StoreChecks(
-                    time=14,
+                    time=Uint64(14),
                     head_slot=Slot(2),
                     head_root_label="block_2",
                 ),
@@ -286,7 +287,7 @@ def test_tick_interval_0_skips_acceptance_when_not_proposer(
             TickStep(
                 interval=15,
                 checks=StoreChecks(
-                    time=15,
+                    time=Uint64(15),
                     head_slot=Slot(2),
                     head_root_label="block_2",
                     attestation_checks=[
@@ -305,7 +306,7 @@ def test_tick_interval_0_skips_acceptance_when_not_proposer(
             TickStep(
                 interval=19,
                 checks=StoreChecks(
-                    time=19,
+                    time=Uint64(19),
                     head_slot=Slot(2),
                     head_root_label="block_2",
                 ),
@@ -342,7 +343,7 @@ def test_tick_interval_0_skips_acceptance_when_not_proposer(
                 interval=20,
                 has_proposal=True,
                 checks=StoreChecks(
-                    time=20,
+                    time=Uint64(20),
                     head_slot=Slot(2),
                     head_root_label="block_2",
                     attestation_checks=[
@@ -361,7 +362,7 @@ def test_tick_interval_0_skips_acceptance_when_not_proposer(
             TickStep(
                 interval=28,
                 checks=StoreChecks(
-                    time=28,
+                    time=Uint64(28),
                     head_slot=Slot(2),
                     head_root_label="block_2",
                 ),
@@ -391,7 +392,7 @@ def test_tick_interval_0_skips_acceptance_when_not_proposer(
             TickStep(
                 interval=29,
                 checks=StoreChecks(
-                    time=29,
+                    time=Uint64(29),
                     head_slot=Slot(2),
                     head_root_label="block_2",
                     attestation_checks=[
